Split buildSrc up into multiple plugins for smaller, simpler components
