+++ build.gradle
@@ -1,6 +1,8 @@
 plugins {
     id "java"
     id "org.openmicroscopy.insight"
+    id "org.openmicroscopy.distribute"
+    id "org.openmicroscopy.publish"
     id "org.openmicroscopy.additional-repositories" version "5.5.0-m5"
 }
