+++ buildSrc/src/main/groovy/org/openmicroscopy/Utils.groovy
@@ -1,0 +1,11 @@
+package org.openmicroscopy
+
+import org.gradle.api.Project
+import org.gradle.api.artifacts.Configuration
+import org.gradle.api.plugins.JavaPlugin
+
+class Utils {
+    static Configuration getRuntimeClasspathConfiguration(Project project) {
+        project.configurations.findByName(JavaPlugin.RUNTIME_CLASSPATH_CONFIGURATION_NAME)
+    }
+}
